requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.31.0
alembic==1.13.1
python-multipart==0.0.9